refactor: split UserStatusDialog.vue along its existing seams
